Cate 부모 id 타입 오류 수정

+++ src/main/java/kr/co/paywith/pw/data/repository/od/cate/Cate.java
@@ -52,7 +52,7 @@ public class Cate {
 	private List<Cate> subCateList;
 
 	@Column(name = "parent_id")
-	private Long parentId;
+	private Integer parentId;
 
 	@Column(name = "use_fl")
 	private Boolean useFl;
